Implement a "subshell" command.
When invoked from a batch file (or from an ongoing REPL session)
starts a "subshell" REPL session. The intent here is to allow
for exploration and experimentation in the middle of a larger run.

For now, subshells are not very useful, as there isn't a good way
to expose the local SAW environment to the shell.